refactor: [app] move globalChat plugin
