+++ src/module/Chat/createMessage.ts
@@ -1,6 +1,6 @@
 import { randomUUID } from "crypto";
 import { ChatMessage, ChatReplyMessage } from "./entity/index.js";
-import type { ChatContext } from "./chatPlugin.js";
+import type { ChatContext } from "../../plugins/modules/chat/global-chat.auto-load.js";
 
 export default function createMessage({ sender, text, replyMessageId }: {
   sender: ChatContext['sender'],
