+++ src/plugins/modules/chat/global-chat.auto-load.ts
@@ -2,14 +2,16 @@ import { type FastifyInstance } from "fastify";
 import {
   CustomWebsocketEvent,
   NotificationAlertEvent,
-} from "../../ws/index.js";
-import createMessage from "./createMessage.js";
-import type { Chat, ChatMessage, ChatReplyMessage } from "./entity/index.js";
-import initializeChat from "./initializeChatNamespace.js";
+} from "../../../ws/index.js";
+import createMessage from "../../../module/Chat/createMessage.js";
+import type { Chat, ChatMessage, ChatReplyMessage } from "../../../module/Chat/entity/index.js";
+import initializeChat from "../../../module/Chat/initializeChatNamespace.js";
 
 export type ChatContext = ReturnType<ReturnType<typeof initializeChat>>;
 
-export default async function chatPlugin(
+export const autoConfig = { path: "/global-chat" };
+
+export default async function(
   fastify: FastifyInstance,
   options: { path: string },
 ) {
